+++ CHANGELOG.md
@@ -27,6 +27,7 @@ You should also include the user name that made the change.
 - Client: Poll highlights in explore page @syuilo
 - Client: Improve deck UI @syuilo
 - Client: Word mute also checks content warnings @Johann150
+- Client: メニューからページをリロードできるように @syuilo
 - ユーザーにモデレーションメモを残せる機能 @syuilo
 - Make possible to delete an account by admin @syuilo
 - Improve player detection in URL preview @mei23